% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehrc_style.R
\name{ehrc_style}
\alias{ehrc_style}
\title{Add EHRC theme to ggplot chart}
\usage{
ehrc_style()
}
\description{
This function allows you to add the ehrc theme to your ggplot graphics.
}
\examples{
library("ggplot2")
library("gapminder")
gap = as.data.frame(gapminder)
line_df = gap[gap$country \%in\% "China", ]
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
  geom_line(colour = "#1380A1", linewidth = 1) +
  geom_hline(yintercept = 0, linewidth = 1, colour="#333333") +
  ehrc_style() +
  labs(title="Living longer",
       subtitle = "Life expectancy in China 1952-2007")
}
\keyword{ehrc_style}
