% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and save EHRC ggplot chart}
\usage{
finalise_plot(
  plot_name,
  source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file("data", package = "ehrcplot"), "icon.png")
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to EHRC block image that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for
publication for EHRC graphic.
It will left align your title, subtitle and source, add the EHRC blocks at the bottom right and save it to your specified location.
}
\examples{
library("ggplot2")
library("gapminder")
gap = as.data.frame(gapminder)
line_df = gap[gap$country \%in\% "China", ]
g <- ggplot(line_df, aes(x = year, y = lifeExp)) +
  geom_line(colour = "#1380A1", linewidth = 1) +
  geom_hline(yintercept = 0, linewidth = 1, colour="#333333") +
  ehrc_style() +
  labs(title="Living longer",
       subtitle = "Life expectancy in China 1952-2007")
finalise_plot(plot_name = g,
source = "Gapminder",
save_filepath = "chinaLifeExpectacy.png",
width_pixels = 640,
height_pixels = 450
)

}
\keyword{finalise_plot}
